Updating file names so that reality matches current Snapshot repo
