+++ src/plugins/safeSnap/components/SafeTransactions.vue
@@ -10,7 +10,7 @@ import { getIpfsUrl, shorten } from '@/helpers/utils';
 
 import SafeSnapTooltip from './Tooltip.vue';
 import SafeSnapHandleOutcome from './HandleOutcome.vue';
-import SafeSnapHandleOutcomeReality from './HandleOutcomeReality.vue';
+import SafeSnapHandleOutcomeUma from './HandleOutcomeUma.vue';
 import SafeSnapFormImportTransactionsButton from './Form/ImportTransactionsButton.vue';
 import SafeSnapFormTransactionBatch from './Form/TransactionBatch.vue';
 
@@ -61,8 +61,8 @@ export default {
   components: {
     SafeSnapTooltip,
     SafeSnapFormImportTransactionsButton,
-    SafeSnapHandleOutcomeReality,
     SafeSnapHandleOutcome,
+    SafeSnapHandleOutcomeUma,
     SafeSnapFormTransactionBatch
   },
   props: [
@@ -256,7 +256,7 @@ export default {
           @import="handleImport($event)"
         />
 
-        <SafeSnapHandleOutcomeReality
+        <SafeSnapHandleOutcome
           v-if="
             preview &&
             proposalResolved &&
@@ -270,7 +270,7 @@ export default {
           :network="transactionConfig.network"
         />
 
-        <SafeSnapHandleOutcome
+        <SafeSnapHandleOutcomeUma
           v-if="
             preview &&
             proposalResolved &&
